requirements.txt
py-cord>=2.5.0
pydantic>=2.8.0
google-genai
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
